Avoided variable translations.

+++ src/games/renpy/__init__.py
@@ -393,7 +393,7 @@ class RenpyFrame(object):
         return True
     def __google_translate(self):
         try:
-            translated_comment = " # translated"
+            translated_comment = "# translated"
             if self.translatewithGoogleTranslate:
                 if len(self.googleTranslateLanguageCode) > 0:
                     if path.exists(self.tlfilesdir):
@@ -403,35 +403,43 @@ class RenpyFrame(object):
                                 if path.isfile(reallocation) and reallocation.endswith(".rpy"):
                                     self.progress = "Translating \""+reallocation+"\"..."
                                     newtexts = ""
+                                    already_translated = False
                                     with open(reallocation, "r+") as tlfile:
                                         file_lines = tlfile.readlines()
-                                        for file_line in file_lines:
-                                            file_line = str(file_line)
-                                            file_line_stripped = file_line.strip()
-                                            if not file_line_stripped.startswith("#") and not file_line.startswith("translate ") and not file_line_stripped.startswith("old") and not len(file_line_stripped) == 0 and not file_line_stripped.endswith(translated_comment):
-                                                p = re.compile('\\"(.*)\\"',)
-                                                result = p.search(file_line)
-                                                if result is not None:
-                                                    translate_text = result.group(1)
-                                                    print("Original: "+translate_text)
-                                                    m = re.findall(r'\[.+?\]', translate_text)
-                                                    variable_map = {}
-                                                    for i in range(len(m)):
-                                                        variable_map["["+str(i)+"]"] = m[i]
-                                                        translate_text = translate_text.replace(m[i], "["+str(i)+"]")
-                                                    translated = translator.translate(translate_text, dest = self.googleTranslateLanguageCode)
-                                                    translated_text = translated.text
-                                                    print("Translated: "+ translated_text)
-                                                    for key, value in variable_map.items():
-                                                        translated_text = translated_text.replace(key, value)
-                                                    print("Restored: "+ translated_text)
-                                                    file_line = file_line.replace("\""+result.group(1)+"\"", "\""+translated_text+"\""+translated_comment)
-                                            newtexts += file_line
+                                        if file_lines[len(file_lines)-1].strip().startswith(translated_comment):
+                                            already_translated = True
+                                        else:
+                                            for file_line in file_lines:
+                                                file_line = str(file_line)
+                                                file_line_stripped = file_line.strip()
+                                                if not file_line_stripped.startswith("#") and not file_line.startswith("translate ") and not file_line_stripped.startswith("old") and not len(file_line_stripped) == 0:
+                                                    p = re.compile('\\"(.*)\\"',)
+                                                    result = p.search(file_line)
+                                                    if result is not None:
+                                                        translate_text = result.group(1)
+                                                        print("Original: "+translate_text)
+                                                        m = re.findall(r'\[.+?\]', translate_text)
+                                                        variable_map = {}
+                                                        for i in range(len(m)):
+                                                            replaced_variable = "__["+str(i)+"]__"
+                                                            variable_map[replaced_variable] = m[i]
+                                                            translate_text = translate_text.replace(m[i], replaced_variable)
+                                                        translated = translator.translate(translate_text, dest = self.googleTranslateLanguageCode)
+                                                        translated_text = translated.text
+                                                        print("Translated: "+ translated_text)
+                                                        for key, value in variable_map.items():
+                                                            translated_text = translated_text.replace(key, value)
+                                                        print("Restored: "+ translated_text)
+                                                        file_line = file_line.replace("\""+result.group(1)+"\"", "\""+translated_text+"\"")
+                                                newtexts += file_line
                                     tlfile.closed
-                                    with open(reallocation, "w") as tlfile2:
-                                        tlfile2.write(newtexts)
-                                    tlfile2.closed
-                                    self.progress = "Translation of \""+reallocation+"\" successfull!" 
+                                    if already_translated:
+                                        self.progress = "Translation of \""+reallocation+"\" skipped. Because it's already translated!" 
+                                    else:
+                                        with open(reallocation, "w") as tlfile2:
+                                            tlfile2.write(newtexts+"\n"+translated_comment)
+                                        tlfile2.closed
+                                        self.progress = "Translation of \""+reallocation+"\" successfull!" 
                         self.progress = "Translation completed! Please launch the game and check if has any error."
                     else:
                         self.progress =  "Translation folder \""+self.tlfilesdir+"\" not found!"
